requirements.txt
absl-py==1.2.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
audioread==2.1.9
autopep8==1.7.0
backcall==0.2.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docutils==0.19
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.22.0
importlib-metadata==4.12.0
ipykernel==6.14.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.8.2
kiwisolver==1.4.3
libclang==14.0.6
librosa==0.9.1
llvmlite==0.38.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplot==0.1.9
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.6
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pkginfo==1.8.3
pooch==1.6.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
pyjion==1.2.6
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.0
pyzmq==23.1.0
readme-renderer==36.0
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==2.0.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
SimpleWebSocketServer==0.1.1
six==1.16.0
snakeviz==2.1.1
SoundFile==0.10.3.post1
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflowjs==3.19.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
twine==4.0.1
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.9
ushlex==0.99.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
